+++ libkeymaster/tlcTeeKeymaster_if.h
@@ -38,6 +38,7 @@ extern "C" {
 
 #include <stdint.h>
 #include <stdbool.h>
+#include <string.h>
 
 
 /**
@@ -149,7 +150,7 @@ teeResult_t TEE_RSAGenerateKeyPair(
     uint32_t            keyDataLength,
     uint32_t            keySize,
     uint32_t            exponent,
-    uint32_t*           soLen);
+    size_t*             soLen);
 
 
 /**
@@ -171,7 +172,7 @@ teeResult_t TEE_RSASign(
     const uint8_t*  plainData,
     const uint32_t  plainDataLength,
     uint8_t*        signatureData,
-    uint32_t*       signatureDataLength,
+    size_t*         signatureDataLength,
     teeRsaSigAlg_t  algorithm);
 
 
@@ -195,7 +196,7 @@ teeResult_t TEE_RSAVerify(
     const uint8_t*  plainData,
     const uint32_t  plainDataLength,
     const uint8_t*  signatureData,
-    const uint32_t  signatureDataLength,
+    const size_t    signatureDataLength,
     teeRsaSigAlg_t  algorithm,
     bool            *validity);
 
@@ -213,7 +214,7 @@ teeResult_t TEE_RSAVerify(
 teeResult_t TEE_HMACKeyGenerate(
     uint8_t*  keyData,
     uint32_t  keyDataLength,
-    uint32_t* soLen);
+    size_t*   soLen);
 
 
 /**
@@ -235,7 +236,7 @@ teeResult_t TEE_HMACSign(
     const uint8_t*  plainData,
     const uint32_t  plainDataLength,
     uint8_t*        signatureData,
-    uint32_t*       signatureDataLength,
+    size_t*         signatureDataLength,
     teeDigest_t     digest);
 
 
@@ -259,7 +260,7 @@ teeResult_t TEE_HMACVerify(
     const uint8_t*  plainData,
     const uint32_t  plainDataLength,
     const uint8_t*  signatureData,
-    const uint32_t  signatureDataLength,
+    const size_t    signatureDataLength,
     teeDigest_t     digest,
     bool            *validity);
 
@@ -293,7 +294,7 @@ teeResult_t TEE_KeyImport(
     const uint8_t*  keyData,
     const uint32_t  keyDataLength,
     uint8_t*        soData,
-    uint32_t*       soDataLength);
+    size_t*         soDataLength);
 
 
 /**
